+++ ImageFeed/Profile/ProfileImageService.swift
@@ -27,75 +27,33 @@ final class ProfileImageService {
         guard let token = OAuth2TokenStorage().token else { return }
         let request = makeRequest(username: username, token: token)
         
-        let task = object(for: request) { [weak self] result in
+        let task = urlSession.objectTask(for: request) { [weak self] (result: Result<UserResult, Error>) in
             guard let self = self else { return }
-            switch result {
-            case .success(let body):
-                let avatarURL = body.profileImage.small
-                self.avatarURL = avatarURL
-                completion(.success(avatarURL))
-            case .failure(let error):
-                completion(.failure(error))
+            DispatchQueue.main.async {
+                switch result {
+                case .success(let body):
+                    let avatarURL = body.profileImage.small
+                    self.avatarURL = avatarURL
+                    completion(.success(avatarURL))
+                    self.task = nil
+                case .failure(let error):
+                    completion(.failure(error))
+                }
             }
         }
+        self.task = task
         task.resume()
-        
     }
-    
+}
+
+extension ProfileImageService {
     private func makeRequest(username: String, token: String) -> URLRequest {
-        var request = makeRequest(
+        var request = URLRequest.makeHTTPRequest(
             path: "/users"
             + "/\(username)",
             httpMethod: "GET",
             baseURL: defaultBaseURL)
         request.setValue("Bearer \(token)", forHTTPHeaderField: "Authorization")
         return request
-    }
-}
-
-extension ProfileImageService {
-    private func makeRequest(
-        path: String,
-        httpMethod: String,
-        baseURL: URL = defaultBaseURL
-    ) -> URLRequest {
-        var request = URLRequest(url: URL(string: path, relativeTo: baseURL)!)
-        request.httpMethod = httpMethod
-        return request
-    }
-}
-
-extension ProfileImageService {
-    private func data(for request: URLRequest, completion: @escaping (Result<Data, Error>) -> Void) -> URLSessionTask {
-        let fulfillCompletionOnTheMainThread: (Result<Data, Error>) -> Void = { result in
-            DispatchQueue.main.async {
-                completion(result)
-            }
-        }
-
-        let task = urlSession.dataTask(with: request, completionHandler: { data, response, error in
-            if let data = data, let response = response, let statusCode = (response as? HTTPURLResponse)?.statusCode {
-                if 200 ..< 300 ~= statusCode {
-                    fulfillCompletionOnTheMainThread(.success(data))
-                } else {
-                    fulfillCompletionOnTheMainThread(.failure(NetworkError.httpStatusCode(statusCode)))
-                }
-            } else if let error = error {
-                fulfillCompletionOnTheMainThread(.failure(NetworkError.urlRequestError(error)))
-            } else {
-                fulfillCompletionOnTheMainThread(.failure(NetworkError.urlSessionError))
-            }
-        })
-        return task
-    }
-    
-    private func object(for request: URLRequest, completion: @escaping (Result<UserResult, Error>) -> Void) -> URLSessionTask {
-        let decoder = JSONDecoder()
-        return data(for: request) { (result: Result<Data, Error>) in
-            let response = result.flatMap { data -> Result<UserResult, Error> in
-                Result { try decoder.decode(UserResult.self, from: data) }
-            }
-            completion(response)
-        }
     }
 }
