+++ ImageFeed/Profile/ProfileService.swift
@@ -26,7 +26,7 @@ struct Profile {
     var loginName: String {
         return "@\(username)"
     }
-    }
+}
 
 final class ProfileService {
     private let urlSession = URLSession.shared
@@ -40,22 +40,29 @@ final class ProfileService {
         task?.cancel()
                 
         let request = makeRequest(token: token)
-        let task = object(for: request) { [weak self] result in
+        
+        let task = urlSession.objectTask(for: request) { [weak self] (result: Result<ProfileResult, Error>) in
             guard let self = self else { return }
-            switch result {
-            case .success(let body):
-                let profile = Profile(username: body.username, firstName: body.firstName, lastName: body.lastName, bio: body.bio ?? "")
-                self.profile = profile
-                completion(.success(profile))
-            case .failure(let error):
-                completion(.failure(error))
+            DispatchQueue.main.async {
+                switch result {
+                case .success(let body):
+                    let profile = Profile(username: body.username, firstName: body.firstName, lastName: body.lastName, bio: body.bio ?? "")
+                        self.profile = profile
+                        completion(.success(profile))
+                    self.task = nil
+                case .failure(let error):
+                    completion(.failure(error))
+                }
             }
         }
+        self.task = task
         task.resume()
     }
-    
+}
+
+extension ProfileService {
     private func makeRequest(token: String) -> URLRequest {
-        var request = makeRequest(
+        var request = URLRequest.makeHTTPRequest(
             path: "/me",
             httpMethod: "GET",
             baseURL: defaultBaseURL)
@@ -63,52 +70,3 @@ final class ProfileService {
         return request
     }
 }
-
-extension ProfileService {
-    private func makeRequest(
-        path: String,
-        httpMethod: String,
-        baseURL: URL = defaultBaseURL
-    ) -> URLRequest {
-        var request = URLRequest(url: URL(string: path, relativeTo: baseURL)!)
-        request.httpMethod = httpMethod
-        return request
-    }
-}
-
-extension ProfileService {
-    private func data(for request: URLRequest, completion: @escaping (Result<Data, Error>) -> Void) -> URLSessionTask {
-        let fulfillCompletionOnTheMainThread: (Result<Data, Error>) -> Void = { result in
-            DispatchQueue.main.async {
-                completion(result)
-            }
-        }
-
-        let task = urlSession.dataTask(with: request, completionHandler: { data, response, error in
-            if let data = data, let response = response, let statusCode = (response as? HTTPURLResponse)?.statusCode {
-                if 200 ..< 300 ~= statusCode {
-                    fulfillCompletionOnTheMainThread(.success(data))
-                } else {
-                    fulfillCompletionOnTheMainThread(.failure(NetworkError.httpStatusCode(statusCode)))
-                }
-            } else if let error = error {
-                fulfillCompletionOnTheMainThread(.failure(NetworkError.urlRequestError(error)))
-            } else {
-                fulfillCompletionOnTheMainThread(.failure(NetworkError.urlSessionError))
-            }
-        })
-        return task
-    }
-    
-    private func object(for request: URLRequest, completion: @escaping (Result<ProfileResult, Error>) -> Void) -> URLSessionTask {
-        let decoder = JSONDecoder()
-        return data(for: request) { (result: Result<Data, Error>) in
-            let response = result.flatMap { data -> Result<ProfileResult, Error> in
-                Result { try decoder.decode(ProfileResult.self, from: data) }
-            }
-            completion(response)
-        }
-    }
-    
-}
-
