+++ ImageFeed/ServiceAccessLayer/OAuth2Service.swift
@@ -1,19 +1,29 @@
 import UIKit
 
-enum NetworkError: Error {
-    case httpStatusCode(Int)
-    case urlRequestError(Error)
-    case urlSessionError
+struct OAuthTokenResponseBody: Codable {
+    let accessToken: String
+    let tokenType: String
+    let scope: String
+    let createdAt: Int
+
+    enum CodingKeys: String, CodingKey {
+        case accessToken = "access_token"
+        case tokenType = "token_type"
+        case scope
+        case createdAt = "created_at"
+    }
 }
 
 final class OAuth2Service {
     static let shared = OAuth2Service()
-
+    private init() {}
+    
     private let urlSession = URLSession.shared
     
     private var task: URLSessionTask?
+    
     private var lastCode: String?
-
+    
     private (set) var authToken: String? {
         get {
             return OAuth2TokenStorage().token
@@ -22,30 +32,39 @@ final class OAuth2Service {
             OAuth2TokenStorage().token = newValue
         }
     }
-
+    
     func fetchOAuthToken(_ code: String, completion: @escaping (Result<String, Error>) -> Void) {
         assert(Thread.isMainThread)
+        
         if lastCode == code { return }
         task?.cancel()
         lastCode = code
         
         let request = makeRequest(code: code)
-        let task = object(for: request) { [weak self] result in
+        
+        let task = urlSession.objectTask(for: request) { [weak self] (result: Result<OAuthTokenResponseBody, Error>) in
             guard let self = self else { return }
-            switch result {
-            case .success(let body):
-                let authToken = body.accessToken
-                self.authToken = authToken
-                completion(.success(authToken))
-            case .failure(let error):
-                completion(.failure(error))
+            DispatchQueue.main.async {
+                switch result {
+                case .success(let body):
+                    let authToken = body.accessToken
+                    self.authToken = authToken
+                    completion(.success(authToken))
+                    self.task = nil
+                case .failure(let error):
+                    completion(.failure(error))
+                    self.lastCode = nil
+                }
             }
         }
+        self.task = task
         task.resume()
-    }
+            }
+        }
 
+extension OAuth2Service {
     private func makeRequest(code: String) -> URLRequest {
-        makeRequest(
+        URLRequest.makeHTTPRequest(
             path: "/oauth/token"
             + "?client_id=\(accessKey)"
             + "&&client_secret=\(secretKey)"
@@ -55,70 +74,5 @@ final class OAuth2Service {
             httpMethod: "POST",
             baseURL: URL(string: "https://unsplash.com")!
         )
-    }
-
-    private struct OAuthTokenResponseBody: Codable {
-        let accessToken: String
-        let tokenType: String
-        let scope: String
-        let createdAt: Int
-
-        enum CodingKeys: String, CodingKey {
-            case accessToken = "access_token"
-            case tokenType = "token_type"
-            case scope
-            case createdAt = "created_at"
-        }
-    }
-}
-
-// MARK: - Network Client
-
-extension OAuth2Service {
-    private func data(for request: URLRequest, completion: @escaping (Result<Data, Error>) -> Void) -> URLSessionTask {
-        let fulfillCompletionOnTheMainThread: (Result<Data, Error>) -> Void = { result in
-            DispatchQueue.main.async {
-                completion(result)
-            }
-        }
-
-        let task = urlSession.dataTask(with: request, completionHandler: { data, response, error in
-            if let data = data, let response = response, let statusCode = (response as? HTTPURLResponse)?.statusCode {
-                if 200 ..< 300 ~= statusCode {
-                    fulfillCompletionOnTheMainThread(.success(data))
-                } else {
-                    fulfillCompletionOnTheMainThread(.failure(NetworkError.httpStatusCode(statusCode)))
-                }
-            } else if let error = error {
-                fulfillCompletionOnTheMainThread(.failure(NetworkError.urlRequestError(error)))
-            } else {
-                fulfillCompletionOnTheMainThread(.failure(NetworkError.urlSessionError))
-            }
-        })
-        return task
-    }
-
-    private func object(for request: URLRequest, completion: @escaping (Result<OAuthTokenResponseBody, Error>) -> Void) -> URLSessionTask {
-        let decoder = JSONDecoder()
-        return data(for: request) { (result: Result<Data, Error>) in
-            let response = result.flatMap { data -> Result<OAuthTokenResponseBody, Error> in
-                Result { try decoder.decode(OAuthTokenResponseBody.self, from: data) }
-            }
-            completion(response)
-        }
-    }
-}
-
-// MARK: - Request Factory
-
-extension OAuth2Service {
-    private func makeRequest(
-        path: String,
-        httpMethod: String,
-        baseURL: URL = defaultBaseURL
-    ) -> URLRequest {
-        var request = URLRequest(url: URL(string: path, relativeTo: baseURL)!)
-        request.httpMethod = httpMethod
-        return request
     }
 }
